refactor: move expand_path / get_project_dirs_up_to_home to utils/paths
These are general path utilities, not config schema concerns. Lifting them
out of src/config/config.cpp puts them next to the rest of the path helpers
in src/utils/paths.{hpp,cpp}. Pure code reorganization — no behavior change.

Callers updated: src/skills/skill_init.cpp and src/daemon/worker.cpp now
include <utils/paths.hpp> directly instead of relying on transitive include
through config.hpp.
